move ProductCard and Pagination in separate components
