+++ frontend/src/components/Pagination.jsx
@@ -1,0 +1,69 @@
+// components/Pagination.jsx
+import React from "react";
+import PropTypes from "prop-types";
+
+export default function Pagination({
+  page,
+  totalPages,
+  pageSize,
+  onPageChange,
+  onPageSizeChange,
+  pageSizeOptions = [12, 24, 48, 96],
+  className = "",
+}) {
+  const canPrev = page > 1;
+  const canNext = page < totalPages;
+
+  return (
+    <div className={`flex items-center justify-between mt-6 ${className}`}>
+      <div className="flex items-center space-x-2">
+        <button
+          type="button"
+          onClick={() => onPageChange(Math.max(1, page - 1))}
+          disabled={!canPrev}
+          className="px-3 py-1 bg-white border rounded-lg disabled:opacity-50"
+          aria-label="Previous page">
+          ← Prev
+        </button>
+
+        <span className="text-sm text-gray-600">
+          Page <strong>{page}</strong> of <strong>{totalPages}</strong>
+        </span>
+
+        <button
+          type="button"
+          onClick={() => onPageChange(Math.min(totalPages, page + 1))}
+          disabled={!canNext}
+          className="px-3 py-1 bg-white border rounded-lg disabled:opacity-50"
+          aria-label="Next page">
+          Next →
+        </button>
+      </div>
+
+      <div className="flex items-center space-x-2">
+        <span className="text-sm text-gray-600">Per page:</span>
+        <select
+          value={pageSize}
+          onChange={(e) => onPageSizeChange(parseInt(e.target.value, 10))}
+          className="px-2 py-1 border rounded-lg"
+          aria-label="Items per page">
+          {pageSizeOptions.map((sz) => (
+            <option key={sz} value={sz}>
+              {sz}
+            </option>
+          ))}
+        </select>
+      </div>
+    </div>
+  );
+}
+
+Pagination.propTypes = {
+  page: PropTypes.number.isRequired,
+  totalPages: PropTypes.number.isRequired,
+  pageSize: PropTypes.number.isRequired,
+  onPageChange: PropTypes.func.isRequired,
+  onPageSizeChange: PropTypes.func.isRequired,
+  pageSizeOptions: PropTypes.arrayOf(PropTypes.number),
+  className: PropTypes.string,
+};
